backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.13.1
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pandas==2.1.3
numpy==1.25.2
PyPDF2==3.0.1
python-docx==1.1.0
lxml==4.9.3
beautifulsoup4==4.12.2
jira==3.5.0
azure-devops==7.1.0b3
cryptography==41.0.7
